requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
black==24.4.2
blis==0.7.11
blosc2==2.6.2
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
cssselect==1.2.0
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.14.0
fsspec==2024.5.0
goose3==3.1.19
h5py==3.11.0
idna==3.7
imageio==2.34.1
inflect==7.2.1
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lazy_loader==0.4
lxml==5.2.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
mypy-extensions==1.0.0
ndindex==1.8
networkx==3.3
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
py-cpuinfo==9.0.0
pyahocorasick==2.1.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
regex==2024.5.15
requests==2.32.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12.1
tables==3.9.2
thinc==8.2.3
threadpoolctl==3.5.0
tifffile==2024.5.22
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typeguard==4.3.0
typer==0.9.4
typing_extensions==4.12.1
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.3
weasel==0.3.4
